requirements.txt
mlabs_messenger==0.1.2
requests==2.18.4
Flask==0.12.2
python-dotenv==0.7.1
messenger==1.0
